\encoding{UTF-8}
\name{mapRDP}
\alias{mapRDP}
\title{Map RDP classifications to NCBI taxonomy}
\description{
Maps RDP classifications to NCBI taxonomy by automatic matching of taxonomic names, with manual mappings for some groups.
}

\usage{
  mapRDP(RDP = NULL)
}

\arguments{
  \item{RDP}{data frame, output from \code{getRDP}}
}

\details{
\code{getmap} maps RDP names to the NCBI taxonomy.
RDP group names are made by combining the rank and name with an underscore separator (e.g. \samp{genus_Escherichia/Shigella}).
Some particular group names are changed to correspond with the NCBI taxonomy as follows:
\tabular{ll}{
    \strong{RDP} \tab \strong{NCBI} \cr
    genus_Escherichia/Shigella \tab genus_Escherichia \cr
    phylum_Cyanobacteria/Chloroplast \tab phylum_Cyanobacteria \cr
    genus_Marinimicrobia_genera_incertae_sedis \tab species_Candidatus Marinimicrobia bacterium \cr
    class_Cyanobacteria \tab phylum_Cyanobacteria \cr
    genus_Spartobacteria_genera_incertae_sedis \tab species_Spartobacteria bacterium LR76 \cr
    class_Planctomycetacia \tab class_Planctomycetia \cr
    class_Actinobacteria \tab phylum_Actinobacteria \cr
    order_Rhizobiales \tab order_Hyphomicrobiales \cr
    genus_Gp1 \tab genus_Acidobacterium \cr
    genus_Gp6 \tab genus_Luteitalea \cr
    genus_GpI \tab genus_Nostoc \cr
    genus_GpIIa \tab genus_Synechococcus \cr
    genus_GpVI \tab genus_Pseudanabaena \cr
    family_Family II \tab family_Synechococcaceae \cr
    genus_Subdivision3_genera_incertae_sedis \tab family_Verrucomicrobia subdivision 3 \cr
    order_Clostridiales \tab order_Eubacteriales \cr
    family_Ruminococcaceae \tab family_Oscillospiraceae
}
The group names are then matched to the NCBI taxa listed in \code{extdata/chem16S/taxon_AA.csv}.
The \code{protein} and \code{organism} columns in this file hold the rank and taxon name extracted from the RefSeq database.
Only exactly matching names are mapped; any unmatching names are excluded from the mapping, and the most abundant unmapped groups are printed.
The return value consists of rownumbers in the data frame generated by reading \code{taxon_AA.csv}.
}

\value{
  Integer vector with length equal to number of rows of RDP.
  Values are indices of taxonomic groups in the NCBI taxonomy (as stored in \code{taxon_AA.csv}), or NA for no matching taxon.
  Attributes \code{unmapped_groups} and \code{unmapped_percent} have the RDP names of unmapped groups and their percentage of the total classification count.
}

\examples{
file <- system.file("extdata/RDP/SMS+12.tab.xz", package = "chem16S")
RDP <- readRDP(file)
map <- mapRDP(RDP)
}
