\encoding{UTF-8}
\name{chem16S-package}
\alias{chem16S-package}
\alias{chem16S}
\docType{package}
\title{Chemical metrics of community reference proteomes from 16S rRNA sequences}

\description{

Functions and data to calculate chemical metrics (carbon oxidation state (\ZC) and stoichiometric hydration state (\nH2O)) for reference proteomes for microbial communities.
Community reference proteomes are generated by combining reference proteomes of taxa (derived from the RefSeq database) with taxonomic classifications of 16S rRNA gene sequences.

}

\details{

  \itemize{
    \item \code{\link{readRDP}} - Read and filter an RDP Classifier table
    \item \code{\link{mapRDP}} - Map RDP Classifier assignments to the NCBI taxonomy
    \item \code{\link{getmetrics}} - Get chemical metrics for community reference proteomes
    \item \code{\link{getmdat_example}} - Example function for retrieving sample metadata
    \item \code{\link{plotmet}} - Plotting metrics with symbols and colors based on metadata
  }

The datasets and computational methods are described by Dick and Tan (2022).

This package is used to make plots for the \samp{geo16S} and \samp{orp16S} papers in the JMDplots package on GitHub (\url{https://github.com/jedick/JMDplots}).

}

\section{Files in extdata/RefSeq}{
  This directory contains two sets of files:
    1) scripts to process source RefSeq sequence files to generate amino acid compositions of species-level reference proteomes and taxonomic names;
    2) script and output for amino acid compositions of higher-level taxa.
  The files are based on RefSeq release 206 of 2021-05-21 (O'Leary et al., 2016).

  \describe{

    \item{\file{README.txt}}{Description of steps to generate reference proteomes of species-level taxa (including downloads and shell commands).}
    \item{\file{gencat.sh}}{Helper script to extract microbial protein records from the RefSeq catalog.}
    \item{\file{protein_refseq.R}}{
      R code to sum the amino acid compositions of all proteins for each bacterial, archaeal, and viral species in the NCBI Reference Sequence database.
      The output file is \file{protein_refseq.csv}.
      The first five columns are: \code{protein} (\dQuote{refseq}), \code{organism} (taxonomic id), \code{ref} (organism name), \code{abbrv} (empty), \code{chains} (number of protein sequences for this organism).
      Columns 6 to 25 have the counts of amino acids.
    }
    \item{\file{taxid_names.R}}{
      R code for processing taxonomic IDs; the output file is \file{taxid_names.csv}.
      The columns are NCBI taxonomic ID (taxid), and names at different taxonomic rank (species, genus, family, order, class, phylum, superkingdom).
    }

    \item{NOTE}{To save space, the two above-mentioned CSV output files are not stored here but are in the \code{extdata/RefSeq} directory of the JMDplots package on GitHub (\url{https://github.com/jedick/JMDplots}.}

    \item{\file{taxon_AA.R}}{Functions to create the files listed below:}
    \item{\file{taxon_AA.csv}}{Average amino acid composition of reference proteomes for all species in each genus and at higher taxonomic ranks.}
    \item{\file{taxon_metrics.csv}}{Chemical metrics (\ZC and \nH2O) for amino acid composition of taxa at genus and higher ranks.}
  }


}


\section{Files in extdata/metadata}{

  \describe{
    \item{\file{BGPF13.csv}}{Metadata for Heart Lake Geyser Basin, Yellowstone (Bowen De León et al., 2012).}
    \item{\file{HLA+16.csv}}{Metadata for the Baltic Sea (Herlemann et al., 2016).}
    \item{\file{SMS+12.csv}}{Metadata for Bison Pool, Yellowstone (Swingley et al., 2012).}
  }

}

\section{Files in extdata/RDP}{

  \describe{
    \item{\file{pipeline.R}}{Pipeline for sequence data processing (uses external programs fastq-dump, vsearch, seqtk, RDP Classifier).}
    \item{\file{BGPF13.tab.xz}}{RDP Classifier output for the 16S rRNA dataset of Heart Lake Geyser Basin.}
    \item{\file{HLA+16.tab.xz}}{RDP Classifier output for the 16S rRNA dataset of the Baltic Sea.}
    \item{\file{SMS+12.tab.xz}}{RDP Classifier output for the 16S rRNA dataset of Bison Pool.}
    \item{\file{bergeyTrainingTree.csv}}{Data from \code{bergeyTrainingTree.xml} in the RDP Classifier package version 2.1.3 (\url{https://sourceforge.net/projects/rdp-classifier/}), used for copy-number adjustment (this option is not active by default).}
  }

}


\references{

Bowen De León K, Gerlach R, Peyton BM, Fields MW. 2013. Archaeal and bacterial communities in three alkaline hot springs in Heart Lake Geyser Basin, Yellowstone National Park. \emph{Frontiers in Microbiology} \bold{4}: 330. \doi{10.3389/fmicb.2013.00330}

Dick JM, Tan J. 2022. Chemical links between redox conditions and estimated community proteomes from 16S rRNA and reference protein sequences. \emph{Microbial Ecology} \doi{10.1007/s00248-022-01988-9}

Herlemann DPR, Lundin D, Andersson AF, Labrenz M, Jürgens K. 2016. Phylogenetic signals of salinity and season in bacterial community composition across the salinity gradient of the Baltic Sea. \emph{Frontiers in Microbiology} \bold{7}: 1883. \doi{10.3389/fmicb.2016.01883}

O'Leary NA et al. 2016. Reference sequence (RefSeq) database at NCBI: current status, taxonomic expansion, and functional annotation. \emph{Nucleic Acids Res.} \bold{44}: D733-D745. \doi{10.1093/nar/gkv1189}

Swingley WD, Meyer-Dombard DR, Shock EL, Alsop EB, Falenski HD, Havig JR, Raymond J. 2012. Coordinating environmental genomics and geochemistry reveals metabolic transitions in a hot spring ecosystem. \emph{PLOS One} \bold{7}(6): e38108. \doi{doi.org/10.1371/journal.pone.0038108}

}
