\encoding{UTF-8}
\name{mapRDP}
\alias{mapRDP}
\title{Map RDP classifications to RefSeq (NCBI) or GTDB taxonomy}
\description{
Maps RDP classifications to a taxonomy by automatic matching of taxonomic names, with manual mappings for some groups in the NCBI taxonomy.
}

\usage{
  mapRDP(RDP = NULL, refdb = "RefSeq", quiet = TRUE)
}

\arguments{
  \item{RDP}{data frame, output from \code{\link{readRDP}}}
  \item{refdb}{character, name of reference database (\samp{RefSeq} or \samp{GTDB})}
  \item{quiet}{logical, suppress printed messages?}
}

\details{
\code{mapRDP} maps RDP names to the RefSeq (NCBI) or GTDB taxonomy.
RDP names are made by combining the taxonomic rank and name with an underscore separator (e.g. \samp{genus_Escherichia/Shigella}).
RDP names are then matched to the taxa listed in \file{taxon_AA.csv.xz} found under \file{extdata/RefSeq} or \file{extdata/GTDB}.
The \code{protein} and \code{organism} columns in this file hold the rank and taxon name extracted from the RefSeq or GTDB database.
Only exactly matching names are automatically mapped.

For mapping to the GTDB taxonomy, the RDP Classifier should be run with the GTDB SSU training set.
Example files are provided under \file{extdata/RDP-GTDB}.

For mapping to the RefSeq (NCBI) taxonomy, some group names are manually mapped as follows:
\tabular{ll}{
    \strong{RDP} \tab \strong{NCBI} \cr
    genus_Escherichia/Shigella \tab genus_Escherichia \cr
    phylum_Cyanobacteria/Chloroplast \tab phylum_Cyanobacteria \cr
    genus_Marinimicrobia_genera_incertae_sedis \tab species_Candidatus Marinimicrobia bacterium \cr
    class_Cyanobacteria \tab phylum_Cyanobacteria \cr
    genus_Spartobacteria_genera_incertae_sedis \tab species_Spartobacteria bacterium LR76 \cr
    class_Planctomycetacia \tab class_Planctomycetia \cr
    class_Actinobacteria \tab phylum_Actinobacteria \cr
    order_Rhizobiales \tab order_Hyphomicrobiales \cr
    genus_Gp1 \tab genus_Acidobacterium \cr
    genus_Gp6 \tab genus_Luteitalea \cr
    genus_GpI \tab genus_Nostoc \cr
    genus_GpIIa \tab genus_Synechococcus \cr
    genus_GpVI \tab genus_Pseudanabaena \cr
    family_Family II \tab family_Synechococcaceae \cr
    genus_Subdivision3_genera_incertae_sedis \tab family_Verrucomicrobia subdivision 3 \cr
    order_Clostridiales \tab order_Eubacteriales \cr
    family_Ruminococcaceae \tab family_Oscillospiraceae
}

Change \code{quiet} to FALSE to enable printing of messages about manual RDP to NCBI mappings, most abundant unmapped groups, and overall percentage of classifications mapped to NCBI taxonomy.
}

\value{
  Integer vector with length equal to number of rows of RDP.
  Values are rownumbers in the data frame generated by reading \code{taxon_AA.csv.xz}, or NA for no matching taxon.
  Attributes \code{unmapped_groups} and \code{unmapped_percent} have the RDP names of unmapped groups and their percentage of the total classification count.
}

\examples{
# Mapping with default RDP training set
file <- system.file("extdata/RDP/SMS+12.tab.xz", package = "chem16S")
# Use quiet = FALSE to print messages
RDP <- readRDP(file, quiet = FALSE)
map <- mapRDP(RDP, quiet = FALSE)

# Mapping with GTDB SSU training set
file <- system.file("extdata/RDP-GTDB/SMS+12.tab.xz", package = "chem16S")
RDP_GTDB <- readRDP(file, quiet = FALSE)
map_GTDB <- mapRDP(RDP_GTDB, refdb = "GTDB", quiet = FALSE)

# Mapping of RDP with default training set to RefSeq (NCBI) taxonomy
sum(attributes(map)$unmapped_percent)
# Mapping of RDP with GTDB SSU training set to GTDB taxonomy
# (This is much smaller than the preceding result)
sum(attributes(map_GTDB)$unmapped_percent)
}
